Looks good, but something fails yet
commands are not considered as responses.  We have to add conditionally
commands as possible responses, because there are commands that don't
expect themselves as valid responses.  For example, WHOIS command
expects a RPL_WHOISEND response as final, and WHOIS must not be
registered as a possible response to WHOIS command.